transfer legacy code
Salvaging an idiosyncratic implementation of the logic from the initial D3 infographic project in order to begin conversion into a more flexible plugin. Note that for now this still depends on the word-break property in CSS at extremely small bounding box sizes, because it doesn't dynamically calculate ways to chop a single word into fragment. Will change that soon...