setting up plugin structure
Basic plugin structure is now set up so the textwrap() function can now
be called as part of a D3 method chain. However, the wrapping is still
not actually running properly, mostly because we need to adjust the
internal functions to allow input arguments for the wrapping
boundaries; will likely allow this to be either a simple JavaScript
object or a D3-selected SVG rect.