check for boundaries
I’ve added an input argument to check for desired wrapping boundaries,
and the plugin just exits if it is not provided. Keep that in mind
– this won’t do anything unless d3.textwrap() has an argument inside
the parentheses!

Also added a comment header reiterating that the plugin wrapper is not
functional yet, just so nobody gets confused if they don’t read the the
end of the README.md file on the project’s front page.