partial extraction of positioning information from D3 rect selections
Here we have the first few steps toward accepting a D3 selection as the
bounding argument to the textwrap() method. It’s already getting a bit
tricky to calculate for groups and transforms and animations. For now
we’re just retrieving using getBoundingClientRect(), and a likely end
result is that we’ll require the bounding rect and the text node to be
placed inside the same <g>, which doesn’t seem like an unreasonable
demand of the developers who might eventually use this tool, right?