no further need for getBoundingClientRect()
Actually we no longer need to use this non-D3 function, switched to
using regular D3 attribute retrieval for all bounds properties.

At this point the bounds validation process sanitizes such that the
excess information about the rect node is removed, and the values are
available as a simple property instead of requiring a getter method
like attr(). I’m on the fence as to whether this is the way to go
– might just add a simple retrieval function to the bounds object or
something.