specify proper parent for foreign object insertion
Big progress here in that the text now gets moved to a foreignObject
correctly when you call textwrap() on it as part of a regular D3 method
chain, and as exhaustively detailed in previous commits you can use
hard coded bounds or a D3 rect selection. Unfortunately the
foreignObjectified HTML div still not showing up because the inserted
div does not get sized properly to match the desired bounding box, and
of course there’s still the large looming matter of refactoring the
span logic to enable Internet Explorer compatibility before this plugin
actually becomes useful to anybody.