remove invalid padding attribute
It’s very tempting to allow a regular CSS padding value, but as far as
I can tell this is actually invalid for SVG. So rather than introducing
esoteric pseudo-code, let’s just assume that developers will be willing
to define bounds that reflect the desiring padding value. That is
admittedly a bit of a pain, so perhaps in the future we’ll rethink this
to allow additional optional input arguments and/or append padding to
the bounds object. This is also likely to be an issue with line spacing
aka css line-height, so let’s think up a reasonable solution that works
for both.