always use verified bounds - now working properly with browsers that support foreignObject
Store the result of verify_bounds() into a separate variable, test that
it returned properly during the initialization check, and then
overwrite the input bounds variable. This ensures that even when the
input comes in the format of a D3 selection, it gets converted to a
simple predictable format that we can use downstream. (This may need to
change depending on animations, transitions, transforms, etc.)

As of this commit, calling d3.textwrap() on a text selection now works
properly in all cases where the browser supports foreignObjects.
(Coming up next: Internet Explorer and spans…)