adding "An Illustrative Example" section to introduction.
For users with a certain level of tech experience, it’s easiest to
explain what this plugin does by showing code samples of the HTML/SVG
markup before and after the conversions. May need to expand this
further to cover the bounds or show exact positions and dimensions, but
then again these may be better solved with a bl.ock.s example.