add developer switch to force wrap function
On line 23 there’s now a variable which can be used to force a wrapping
method for development purposes – to test span rendering on a
non-Windows computer, for example. Set the value of the
force_wrap_method variable to the strings ‘spans’ or ‘foreignobjects’
to override the usual browser-driven method of selecting a wrap method.