don't check typeof when deferring
The plugin will defer to whatever else you might define in your code as
d3.selection.prototype.textwrap. Previously, however, it was checking
typeof to make sure that was a function. Removed this check so now
it’ll still exit and play nice even if you’ve assigned that as a
regular non-function property for whatever reason rather than as a
method.