properly translate line-height in tspan rendering
returning computed styles for line height always gives you a px value,
which is right at home in SVG, but it comes with a ‘px’ suffix
appended. Running .replace() clears this out and makes all line-height
adjustments made with CSS work, even if they’re in em or percent values.