add optional configurable padding argument
Many use cases exist where you’ll want to wrap to a defined rect
element but not have the text run edge to edge. The .textwrap()
function now accepts an optional second argument, in addition to the
bounds, which specifies a padding value to be subtracted from the
bounds. This must be provided as a pixel value because we don’t yet
have a way to compute the pixel values of other measurements.