fix associative array arguments
Validating the bounds argument was failing because bounds.x is false if
there’s no x value. Instead of a simple false test we’ll specifically
test with typeof to see if it’s undefined. This means 0 becomes a valid
value. Also forcing a reset on force_wrap_method in cases where it’s
undefined so that the whole plugin doesn’t die in cases where someone
fiddles with the developer flags and forgets to reset them.