Return modified text node after tspans have been inserted
Previously we were not yet actually returning anything from the text
wrapping routine, which means that it wasn’t yet possible to chain
methods. That’s one of the main benefits of this plugin! From here
forward we’re returning the modified text nodes after the text has been
split into tspans. Unfortunately I still haven’t figured out how to do
this for the foreign objects, because the node misreports the
parentNode for <text> as being the root <html> tag, with no mention of
upstream <div>, <svg>, or <g> elements. Hope to solve that next.